Add score to the fhog_svm_det function for more robust detection.

+++ dlib/dlib/fhog_svm_detector.h
@@ -7,8 +7,9 @@
 
 extern bool isfile(const char* str);
 
-// Return [0, 0, 0, 0] if not detected, [left, top, right, bottom] if detected.
-extern const int* fhog_svm_det(const char* img_path, const char* model_path, int length);
+// Return [0, 0, 0, 0, 0.0] if not detected, [left, top, right, bottom, score] if detected.
+// score is a signed float
+extern const double* fhog_svm_det(const char* img_path, const char* model_path, int length);
 
 extern int test_dlib_add(int x, int y);
